+++ packages/riverpod_analyzer_utils/lib/src/riverpod_ast/ref_invocation.dart
@@ -21,8 +21,7 @@ abstract class RefInvocation extends RiverpodAst
     final functionOwner = function.staticElement
         .cast<MethodElement>()
         ?.declaration
-        // ignore: deprecated_member_use, necessary to support older versions of analyzer
-        .enclosingElement;
+        .enclosingElement3;
 
     if (functionOwner == null ||
         // Since Ref is sealed, checking that the function is from the package:riverpod