+++ packages/riverpod_analyzer_utils/lib/src/riverpod_ast/widget_ref_invocation.dart
@@ -26,8 +26,7 @@ abstract class WidgetRefInvocation extends RiverpodAst
     final functionOwner = function.staticElement
         .cast<MethodElement>()
         ?.declaration
-        // ignore: deprecated_member_use, necessary to support older versions of analyzer
-        .enclosingElement;
+        .enclosingElement3;
 
     if (functionOwner == null ||
         // Since Ref is sealed, checking that the function is from the package:riverpod